Change import paths of moved files.
