+++ src/advanced-search/state/actions.js
@@ -1,5 +1,5 @@
-import { fetchData } from '../utils';
-import apiUrls from '../apiUrls';
+import fetchData from '../../utils/fetchData';
+import apiUrls from '../utils/apiUrls';
 
 export const SELECT_FIELD = 'SELECT_FIELD';
 export const UPDATE_INPUT_VALUE = 'UPDATE_INPUT_VALUE';
